# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "电商后台管理系统 API",
        "title": "E-Commerce Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/push/test": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "向当前用户发送一条测试推送通知",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送管理"
                ],
                "summary": "发送测试推送",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_internal_usecase_push.SendTestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/push/token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "注册设备推送通知 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送管理"
                ],
                "summary": "注册推送 Token",
                "parameters": [
                    {
                        "description": "Token 注册请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_internal_usecase_push.RegisterTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_internal_usecase_push.RegisterTokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除设备推送通知 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送管理"
                ],
                "summary": "删除推送 Token",
                "parameters": [
                    {
                        "description": "Token 删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_internal_usecase_push.DeleteTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EricStone1900_ecommerce-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_EricStone1900_ecommerce-backend_internal_usecase_push.DeleteTokenRequest": {
            "type": "object",
            "properties": {
                "device_token": {
                    "type": "string"
                }
            }
        },
        "github_com_EricStone1900_ecommerce-backend_internal_usecase_push.RegisterTokenRequest": {
            "type": "object",
            "properties": {
                "device_token": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                }
            }
        },
        "github_com_EricStone1900_ecommerce-backend_internal_usecase_push.RegisterTokenResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_EricStone1900_ecommerce-backend_internal_usecase_push.SendTestResponse": {
            "type": "object",
            "properties": {
                "sent": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_EricStone1900_ecommerce-backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}